Fix: Proper response and error handling fixed

+++ src/services/member.service.ts
@@ -2,7 +2,6 @@ import { Member } from "../models";
 import { memberRepository } from "../repositories/member.repositories";
 
 export const memberService = {
-  
   async addUser(group_id: number, member_id: number, admin_id: number) {
     try {
       const data = {
@@ -19,12 +18,16 @@ export const memberService = {
     }
   },
 
-  async removeUser(user_id:number){
+  async removeUser(user_id: number) {
     try {
-      return await memberRepository.removeUser(user_id)
+      const userData = await memberRepository.getUser(user_id);
+      if (userData) {
+        return await memberRepository.removeUser(user_id);
+      }else{
+        return false
+      }
     } catch (error) {
-      throw new Error("Error while remove user from group")
-      
+      throw new Error("Error while remove user from group");
     }
-  }
+  },
 };
